Add files via upload
Our final project is the Bubble Shooter game, in which the player aims to clear the field of bubbles by matching 3 or more bubbles of the same color. Bubbles that are not in contact with other bubbles will also be removed. The game is over once the bubbles reach the bottom red line. We set the game to be more challenging as the player shoots more bubbles to the field. This game can be run through Processing.